Restructured project: Model is now a subpackage of ELME
