+++ app/src/main/java/ELME/Model/InputPort.java
@@ -1,16 +1,16 @@
-package Model;
+package ELME.Model;
 
 import java.util.Optional;
 
 /**
- * A kind of {@link Model.Port Port} which is used as inputs for
- * {@link Model.Node Nodes}. This kind of port can only be connected to one
- * other port ({@link Model.OutputPort OutputPort}). Notice that InputPort
+ * A kind of {@link ELME.Model.Port Port} which is used as inputs for
+ * {@link ELME.Model.Node Nodes}. This kind of port can only be connected to one
+ * other port ({@link ELME.Model.OutputPort OutputPort}). Notice that InputPort
  * itself does not store the input value, therefore it must get it from the port
  * it is connected to.
  *
  * Usually it is the InputPort's job to initiate a connection, or to disconnect
- * itself from an other port. {@link Model.OutputPort OutputPorts} act passive
+ * itself from an other port. {@link ELME.Model.OutputPort OutputPorts} act passive
  * in this regard.
  *
  * @author Máté Visnyár
@@ -54,13 +54,13 @@ public class InputPort extends Port {
     }
 
     /**
-     * Connects to supplied {@link Model.OutputPort OutputPort}.
+     * Connects to supplied {@link ELME.Model.OutputPort OutputPort}.
      *
      * Note that since an InputPort can only be connected to one
-     * {@link Model.OutputPort OutputPort} at a time, calling this method
+     * {@link ELME.Model.OutputPort OutputPort} at a time, calling this method
      * disconnects from the previously connected port.
      *
-     * @param port {@link Model.OutputPort OutputPort} to connect to.
+     * @param port {@link ELME.Model.OutputPort OutputPort} to connect to.
      */
     public void connect(OutputPort port) {
         this.connectedPort = port;
