+++ app/src/main/java/ELME/Model/Nodes/ANDNode.java
@@ -1,26 +1,25 @@
-package Model.Nodes;
+package ELME.Model.Nodes;
 
 import java.util.Optional;
 
 /**
- * ODDNode is the implementation of the 2nd type XORNode, where the node is true
- * when the number of true inputs is odd.
+ * Implements the logical AND gate.
  *
  * @author Viktor Bicskei
+ *
  */
-public class ODDNode extends Model.Node {
+public class ANDNode extends ELME.Model.Node {
 
-    public ODDNode() {
-        super("ODD", 3, 1);
+    public ANDNode() {
+        super("AND", 2, 1);
     }
 
     public void evaluateImpl() {
         // SAFETY: It is safe to call .get() because prior to calling this method inputs are tested.
-        if (inputs.stream().filter(x -> x.getValue().get() == true).count() % 2 == 0) {
+        if (inputs.stream().allMatch(x -> x.getValue().get() == true)) {
             setAllOutputs(Optional.of(true));
         } else {
             setAllOutputs(Optional.of(false));
         }
-
     }
 }
