+++ app/src/main/java/ELME/Model/Nodes/ORNode.java
@@ -1,25 +1,26 @@
-package Model.Nodes;
+package ELME.Model.Nodes;
 
 import java.util.Optional;
 
 /**
- * Implements the logical AND gate.
+ * Implements the logical OR gate.
  *
  * @author Viktor Bicskei
  *
  */
-public class ANDNode extends Model.Node {
+public class ORNode extends ELME.Model.Node {
 
-    public ANDNode() {
-        super("AND", 2, 1);
+    public ORNode() {
+        super("OR", 2, 1);
     }
 
     public void evaluateImpl() {
         // SAFETY: It is safe to call .get() because prior to calling this method inputs are tested.
-        if (inputs.stream().allMatch(x -> x.getValue().get() == true)) {
+        if (inputs.stream().anyMatch(x -> x.getValue().get() == true)) {
             setAllOutputs(Optional.of(true));
         } else {
             setAllOutputs(Optional.of(false));
         }
+
     }
 }
